小修了下代码

+++ Engine/Inc/Gear/Core/DX/Resource/UploadHeap.h
@@ -12,7 +12,7 @@ class UploadHeap :public Resource
 {
 public:
 
-	UploadHeap(const UINT size, const D3D12_HEAP_FLAGS flags = D3D12_HEAP_FLAG_NONE);
+	UploadHeap(const UINT64 size, const D3D12_HEAP_FLAGS flags = D3D12_HEAP_FLAG_NONE);
 
 	UploadHeap(UploadHeap&);
 
@@ -22,7 +22,7 @@ public:
 
 	UploadHeap(const UploadHeap&) = delete;
 
-	void update(const void* const data, const UINT size) const;
+	void update(const void* const data, const UINT64 size) const;
 
 	void updateGlobalStates() override;
